9056A Robot's Program
This is program of Team 9056A Starbucks Robot.
The latest version has shown on VEX World 2017.

The most advanced technology that we use is the dynamic power control for moving and lifting which allows robot to get points in a considerable speed. I program to calculate encoder speed in time in order to decrease the voltage at low speed, minimizing the current which enhances power efficiency.
In additional, the use of control curves allows drivers to control robot more accurately. I designed a non-linear function between rocker value and power that the graphic is steeper when rocker value is larger, like a quadratic function (but not).